Pr/9 - updated PR#9 (#10)
* removed code for cron.absent

cron.absent code contain dependency of cadf_packages. cadf_packages will not be added if dispatcher or listener not enabled and without dispatcher and listener config cron is useless. So it will not add cron if its not defined in config and cron not required where where application not installed. It's creating error where dispatcher and listener disabled.

* added logrorate policy and removed cron.absent which creating error

* add logrotate.yml and removed mentioned files.